Refactor: move grpcclient to ecs-agent/
