+++ ecs-agent/grpcclient/credentialsfetcher/credentialsfetcher.proto
